Fix non-rendering compose menu
Currently clicking on the 'compose' button won't take you to
the compose menu, but rather will throw an error saying that the
innerHTML of a null element can't be defined. This change fixes that
by ensuring that the menu is pre-rendered, and only displayed when the
compose button is clicked.

Change-Id: Id85c4673973cd4bb0bdc25110b10007f3650c266
Closes-bug: #1638736