Added Error Response
Added Formating Error response and Confirmation Messages.
When an operation is successfull, its better to send confirmation
message in a standard format. Also when Errors are at Valence-PODM
connection it should be handled and returned as json as per schema
referred in openstack

Closes-Bug: #1643123

Added Error Response

Added Formating Error response and Confirmation Messages.
When an operation is successfull, its better to send confirmation
message in a standard format. Also when Errors are at Valence-PODM
connection it should be handled and returned as json as per schema
referred in openstack

Change-Id: Id539a9b20e289620907d3dc42dfe3104416871a6
Closes-Bug: #1643123