(DO NOT MERGE) Base Ironic Plugin
This is a commit to show what a valence plugin for Ironic
might look like. It is contingent upon the redfish driver
in Ironic being merged https://review.openstack.org/#/c/184653.
In its current basic state, it can set up an Ironic client,
authenticate with keystone, and make a call to enroll one of
the valence controlled systems into Ironic so that Ironic has
control over it.

Change-Id: I010509387a5a6a4ec044b07746f6a723a513aaf5