Fix execute db_manager init failed
Previously, `db_manager init` command failed because pbr generated
script will not pass sys.argv as argument to main function. So fix
it.

And improve etcd_db.init_etcd_db to make it check whether each
directory already exists in etcd at first. If it exist and it's not
directory, delete and re-create it.

Change-Id: I7d9980e0a1b38d988fe8144d0cbd618ac12f5498
Closes-Bug: #1654638