Update requirements manually
The global requirments sync for this repo fails since it lists
packages that are not in the global list. Manually sync
requirements.

Failure log from requirements sync:
http://logs.openstack.org/d8/d89aad257d43fe02e42e44b2a0187dbd9f505c36/post/propose-update-requirements/15a7777/job-output.txt.gz#_2018-01-31_06_51_56_655451

Errors are:
'aniso8601' is not in global-requirements.txt or blacklist.txt
'click' is not in global-requirements.txt or blacklist.txt
'Flask-Cors' is not in global-requirements.txt or blacklist.txt
'itsdangerous' is not in global-requirements.txt or blacklist.txt
'MarkupSafe' is not in global-requirements.txt or blacklist.txt

Change-Id: I1cc19066f2127c37b7122e0068037d41f60fdaaf